Rename mob-authors to new main function.
