+++ src/build-co-author-groups.spec.js
@@ -1,6 +1,6 @@
 const { workspace } = require("../__mocks__/vscode");
 const commands = require("./git/commands");
-const { buildCoAuthorGroups } = require("./mob-authors");
+const { buildCoAuthorGroups } = require("./build-co-author-groups");
 const { Author } = require("./co-author-tree-provider/author");
 const { CoAuthor } = require("./co-author-tree-provider/co-authors");
 const {
